app/requirements.txt
Flask-Restless==0.16.0
Flask-JWT==0.3.2
Flask-Security==1.7.4
Flask-Admin==1.3.0
Flask-SQLAlchemy==2.1
Flask-Login<0.3
Flask-Mail
psycopg2
